requirements.txt
grpcio==1.48.2
grpcio-tools==1.48.2
protobuf==3.20.3

confluent-kafka>=2.3.0

etcd3>=0.12.0

gradio>=4.0.0
pandas